update dwt to use 64-bit tick counts (depends on Vista or newer)

+++ dwt/include/dwt/widgets/Notification.h
@@ -104,7 +104,7 @@ private:
 	NOTIFYICONDATA makeNID() const;
 
 	/// Last tick that tip was updated
-	DWORD lastTick;
+	ULONGLONG lastTick;
 	bool trayHandler(const MSG& msg);
 	bool redisplay();
